wifi: mac80211: Add more NULL guards for sdata in netdev debugfs.
These seem to contain NULL ieee80211_local pointers, which seems wrong.
As with the previous issue in if_read_sdata, this almost certainly stems
from another issue, likely from the mtk patch set. However, we can guard
this case for now, as it is very similar to the other that was already
fixed.

Signed-off-by: Rory Little <rory@candelatech.com>